rfac: Less modules
