+++ cmd/ojster/main_test.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package cli
+package main
 
 import (
 	"bytes"
@@ -31,18 +31,18 @@ func tmpFilePath(t *testing.T, name string) string {
 	return filepath.Join(td, name)
 }
 
-// ----------------------------- Entrypoint basic behavior -----------------------------
+// ----------------------------- entrypoint basic behavior -----------------------------
 
 // TestEntrypoint_NoArgs prints header and returns 0.
 func TestEntrypoint_NoArgs(t *testing.T) {
 	var out, errb bytes.Buffer
 	hdr := "HDR"
-	code := Entrypoint("ojster", []string{}, "v1.2.3", hdr, &out, &errb)
+	code := entrypoint("ojster", []string{}, "v1.2.3", hdr, &out, &errb)
 	if code != 0 {
-		t.Fatalf("Entrypoint returned code %d; want 0", code)
+		t.Fatalf("entrypoint returned code %d; want 0", code)
 	}
 	if out.String() != hdr {
-		t.Fatalf("Entrypoint output mismatch; got %q", out.String())
+		t.Fatalf("entrypoint output mismatch; got %q", out.String())
 	}
 	if errb.Len() != 0 {
 		t.Fatalf("expected no stderr; got %q", errb.String())
@@ -53,9 +53,9 @@ func TestEntrypoint_NoArgs(t *testing.T) {
 func TestEntrypoint_Help(t *testing.T) {
 	var out, errb bytes.Buffer
 	hdr := "HDR"
-	code := Entrypoint("ojster", []string{"help"}, "v", hdr, &out, &errb)
+	code := entrypoint("ojster", []string{"help"}, "v", hdr, &out, &errb)
 	if code != 0 {
-		t.Fatalf("Entrypoint(help) returned %d; want 0", code)
+		t.Fatalf("entrypoint(help) returned %d; want 0", code)
 	}
 	if out.String() != hdr {
 		t.Fatalf("help output mismatch; got %q", out.String())
@@ -66,9 +66,9 @@ func TestEntrypoint_Help(t *testing.T) {
 func TestEntrypoint_Version(t *testing.T) {
 	var out, errb bytes.Buffer
 	hdr := "HDR"
-	code := Entrypoint("ojster", []string{"version"}, "VER-XYZ", hdr, &out, &errb)
+	code := entrypoint("ojster", []string{"version"}, "VER-XYZ", hdr, &out, &errb)
 	if code != 0 {
-		t.Fatalf("Entrypoint(version) returned %d; want 0", code)
+		t.Fatalf("entrypoint(version) returned %d; want 0", code)
 	}
 	if !strings.Contains(out.String(), "VER-XYZ") {
 		t.Fatalf("version output missing; got %q", out.String())
@@ -79,9 +79,9 @@ func TestEntrypoint_Version(t *testing.T) {
 func TestEntrypoint_Unknown(t *testing.T) {
 	var out, errb bytes.Buffer
 	hdr := "HDR"
-	code := Entrypoint("ojster", []string{"nope"}, "v", hdr, &out, &errb)
+	code := entrypoint("ojster", []string{"nope"}, "v", hdr, &out, &errb)
 	if code != 1 {
-		t.Fatalf("Entrypoint(unknown) returned %d; want 1", code)
+		t.Fatalf("entrypoint(unknown) returned %d; want 1", code)
 	}
 	if out.String() != hdr {
 		t.Fatalf("expected header on stdout; got %q", out.String())
@@ -153,9 +153,9 @@ func TestHandleServe_Help(t *testing.T) {
 	}
 }
 
-// ----------------------------- Entrypoint dispatch tests -----------------------------
+// ----------------------------- entrypoint dispatch tests -----------------------------
 
-// TestEntrypoint_SubcommandDispatch verifies Entrypoint dispatches to the correct handlers.
+// TestEntrypoint_SubcommandDispatch verifies entrypoint dispatches to the correct handlers.
 // For most handlers we use the -h help flag so the handler returns quickly without side effects.
 func TestEntrypoint_SubcommandDispatch(t *testing.T) {
 	hdr := "HDR"
@@ -180,7 +180,7 @@ func TestEntrypoint_SubcommandDispatch(t *testing.T) {
 			prog:            "ojster",
 			args:            []string{"version"},
 			wantCode:        0,
-			wantOutContains: "v1.2.3", // Entrypoint prints the version string
+			wantOutContains: "v1.2.3", // entrypoint prints the version string
 		},
 		{
 			name:            "keypair help",
@@ -229,9 +229,9 @@ func TestEntrypoint_SubcommandDispatch(t *testing.T) {
 	for _, c := range cases {
 		t.Run(c.name, func(t *testing.T) {
 			var out, errb bytes.Buffer
-			code := Entrypoint(c.prog, c.args, "v1.2.3", hdr, &out, &errb)
+			code := entrypoint(c.prog, c.args, "v1.2.3", hdr, &out, &errb)
 			if code != c.wantCode {
-				t.Fatalf("Entrypoint(%s %v) returned code %d; want %d; stderr=%q", c.prog, c.args, code, c.wantCode, errb.String())
+				t.Fatalf("entrypoint(%s %v) returned code %d; want %d; stderr=%q", c.prog, c.args, code, c.wantCode, errb.String())
 			}
 			if c.wantOutContains != "" && !strings.Contains(out.String(), c.wantOutContains) {
 				t.Fatalf("stdout did not contain %q; got stdout=%q stderr=%q", c.wantOutContains, out.String(), errb.String())
@@ -265,10 +265,10 @@ func TestEntrypoint_SubcommandFlagParseErrors(t *testing.T) {
 	for _, c := range cases {
 		t.Run(c.name, func(t *testing.T) {
 			var out, errb bytes.Buffer
-			// Entrypoint expects args without the program name; pass as-is so subcommand is first arg.
-			code := Entrypoint("ojster", c.args, "v", hdr, &out, &errb)
+			// entrypoint expects args without the program name; pass as-is so subcommand is first arg.
+			code := entrypoint("ojster", c.args, "v", hdr, &out, &errb)
 			if code != c.wantCode {
-				t.Fatalf("Entrypoint(%v) returned %d; want %d; stderr=%q", c.args, code, c.wantCode, errb.String())
+				t.Fatalf("entrypoint(%v) returned %d; want %d; stderr=%q", c.args, code, c.wantCode, errb.String())
 			}
 			if !strings.Contains(errb.String(), c.wantSubstr) {
 				t.Fatalf("stderr did not contain %q; got: %q", c.wantSubstr, errb.String())
@@ -279,7 +279,7 @@ func TestEntrypoint_SubcommandFlagParseErrors(t *testing.T) {
 
 // ----------------------------- delegation coverage for keypair -----------------------------
 
-// TestEntrypoint_Keypair_Delegation ensures Entrypoint delegates to handleKeypair which calls pqc.KeypairWithPaths.
+// TestEntrypoint_Keypair_Delegation ensures entrypoint delegates to handleKeypair which calls pqc.KeypairWithPaths.
 // This test uses explicit --priv-file and --pub-file flags to avoid writing to default locations.
 func TestEntrypoint_Keypair_Delegation(t *testing.T) {
 	hdr := "HDR"
@@ -290,9 +290,9 @@ func TestEntrypoint_Keypair_Delegation(t *testing.T) {
 	var out, errb bytes.Buffer
 	// pass flags to set output paths
 	args := []string{"keypair", "--priv-file", priv, "--pub-file", pub}
-	code := Entrypoint("ojster", args, "v", hdr, &out, &errb)
+	code := entrypoint("ojster", args, "v", hdr, &out, &errb)
 	if code != 0 {
-		t.Fatalf("Entrypoint(keypair) returned %d; want 0; stderr=%q", code, errb.String())
+		t.Fatalf("entrypoint(keypair) returned %d; want 0; stderr=%q", code, errb.String())
 	}
 
 	// private and public files should exist
@@ -309,33 +309,33 @@ func TestEntrypoint_Keypair_Delegation(t *testing.T) {
 		t.Fatalf("read pub: %v", err)
 	}
 	if !strings.Contains(out.String(), strings.TrimSpace(string(pubB64))) {
-		t.Fatalf("expected Entrypoint output to include public key base64; got stdout=%q stderr=%q", out.String(), errb.String())
+		t.Fatalf("expected entrypoint output to include public key base64; got stdout=%q stderr=%q", out.String(), errb.String())
 	}
 }
 
 // ----------------------------- seal/unseal/run delegation smoke checks -----------------------------
 
-// TestEntrypoint_Seal_MissingPositional ensures Entrypoint dispatches to handleSeal and that
+// TestEntrypoint_Seal_MissingPositional ensures entrypoint dispatches to handleSeal and that
 // handleSeal returns the expected error when the required positional KEY is missing.
 func TestEntrypoint_Seal_MissingPositional(t *testing.T) {
 	hdr := "HDR"
 	var out, errb bytes.Buffer
-	code := Entrypoint("ojster", []string{"seal"}, "v", hdr, &out, &errb)
+	code := entrypoint("ojster", []string{"seal"}, "v", hdr, &out, &errb)
 	if code != 1 {
-		t.Fatalf("Entrypoint(seal missing arg) returned %d; want 1; stdout=%q stderr=%q", code, out.String(), errb.String())
+		t.Fatalf("entrypoint(seal missing arg) returned %d; want 1; stdout=%q stderr=%q", code, out.String(), errb.String())
 	}
 	if !strings.Contains(errb.String(), "seal requires exactly one positional argument: KEY") {
 		t.Fatalf("expected seal missing-arg message; got stderr=%q stdout=%q", errb.String(), out.String())
 	}
 }
 
-// TestEntrypoint_Unseal_Delegation ensures Entrypoint delegates to handleUnseal (help path).
+// TestEntrypoint_Unseal_Delegation ensures entrypoint delegates to handleUnseal (help path).
 func TestEntrypoint_Unseal_Delegation(t *testing.T) {
 	hdr := "HDR"
 	var out, errb bytes.Buffer
-	code := Entrypoint("ojster", []string{"unseal", "-h"}, "v", hdr, &out, &errb)
+	code := entrypoint("ojster", []string{"unseal", "-h"}, "v", hdr, &out, &errb)
 	if code != 0 {
-		t.Fatalf("Entrypoint(unseal -h) returned %d; want 0; stdout=%q stderr=%q", code, out.String(), errb.String())
+		t.Fatalf("entrypoint(unseal -h) returned %d; want 0; stdout=%q stderr=%q", code, out.String(), errb.String())
 	}
 	if !strings.Contains(out.String(), "Usage: ojster unseal") {
 		t.Fatalf("expected unseal usage; got stdout=%q stderr=%q", out.String(), errb.String())
