--- internal/testutil/testutil.go
@@ -1,43 +1,0 @@
-// Copyright 2026 Jip de Beer (Jip-Hop) and ojster contributers
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     https://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-package testutil
-
-import (
-	"encoding/json"
-	"net/http/httptest"
-	"testing"
-)
-
-//
-// ─────────────────────────────────────────────────────────────
-//   TEST HELPERS
-// ─────────────────────────────────────────────────────────────
-//
-
-func ExpectStatus(t *testing.T, rec *httptest.ResponseRecorder, want int) {
-	t.Helper()
-	if rec.Code != want {
-		t.Fatalf("expected %d, got %d (%s)", want, rec.Code, rec.Body.String())
-	}
-}
-
-func DecodeJSON[T any](t *testing.T, data []byte) T {
-	t.Helper()
-	var v T
-	if err := json.Unmarshal(data, &v); err != nil {
-		t.Fatalf("invalid JSON: %v", err)
-	}
-	return v
-}
